chore: declare  and  as module
