+++ go.work
@@ -15,6 +15,8 @@ use (
 	./pkg/build
 	./pkg/build/wire // skip:golangci-lint
 	./pkg/codegen
+	./pkg/infra/log
+	./pkg/infra/tracing
 	./pkg/plugins/codegen
 	./pkg/promlib
 	./pkg/semconv
