+++ pkg/apiserver/go.mod
@@ -5,7 +5,7 @@ go 1.23.7
 require (
 	github.com/google/go-cmp v0.7.0
 	github.com/grafana/authlib/types v0.0.0-20250224151205-5ef97131cc82
-	github.com/grafana/grafana-app-sdk/logging v0.30.0
+	github.com/grafana/grafana-app-sdk/logging v0.32.1
 	github.com/grafana/grafana/pkg/apimachinery v0.0.0-20240701135906-559738ce6ae1
 	github.com/prometheus/client_golang v1.21.0
 	github.com/stretchr/testify v1.10.0
