+++ pkg/infra/log/go.mod
@@ -1,0 +1,27 @@
+module github.com/grafana/grafana/pkg/infra/log
+
+go 1.23.7
+
+require (
+	github.com/go-kit/log v0.2.1
+	github.com/go-stack/stack v1.8.1
+	github.com/grafana/grafana-app-sdk/logging v0.32.1
+	github.com/mattn/go-isatty v0.0.20
+	github.com/stretchr/testify v1.10.0
+	github.com/tjhop/slog-gokit v0.1.4
+	gopkg.in/ini.v1 v1.67.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
+	github.com/go-logfmt/logfmt v0.6.0 // indirect
+	github.com/google/go-cmp v0.7.0 // indirect
+	github.com/kr/pretty v0.3.1 // indirect
+	github.com/pmezard/go-difflib v1.0.1-0.20181226105442-5d4384ee4fb2 // indirect
+	github.com/rogpeppe/go-internal v1.13.1 // indirect
+	go.opentelemetry.io/otel v1.34.0 // indirect
+	go.opentelemetry.io/otel/trace v1.34.0 // indirect
+	golang.org/x/sys v0.30.0 // indirect
+	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
